🔄 chore(package.json): update dependencies and devDependencies in package.json 📦 feat(admin): add new page for listing products and remove old product page The package.json file has been updated to include new dependencies like "@emotion/react", "@emotion/styled", "@mui/material", and "react-icons". These additions enhance the functionality and styling of the application. Additionally, new pages have been added for listing products and adding products, while the old product page has been removed for better organization and clarity in the admin section of the application.
🎨 style(navbar.tsx): update navbar styles and icons for better UI consistency
🔧 chore(globals.css): add active state styles for menu items for improved UX

The changes in navbar.tsx include updating the styles and icons in the navbar component to enhance the user interface consistency and improve the overall look and feel of the application. Additionally, the globals.css file was modified to add active state styles for menu items to provide better visual feedback to users when interacting with the menu.