🔨 refactor(page.tsx): remove unused handleOpenModal function 🚀 feat(page.tsx): set default value for bankSelected2 state and update it based on editPenitip The handleOpenModal function was removed as it was no longer being used in the component. The bankSelected2 state now has a default value of 'BCA' and is updated based on the editPenitip data, improving the initialization and handling of the bank selection in the component.
🔧 fix(page.tsx): add Dialog component import to support modal functionality
🚀 feat(page.tsx): implement modal functionality for editing titipan data
The Dialog component import from '@headlessui/react' was added to support the modal functionality. Modal functionality was implemented to allow users to edit titipan data in a more user-friendly and interactive way. The modal opens when the "Edit" button is clicked, displaying the necessary information for editing the titipan data.

🚀 feat(admin): add modal dialog for editing titipan details
The commit adds a modal dialog component to the TambahTitipan page for editing
the details of a titipan. This allows users to view and modify titipan
information in a more user-friendly and interactive way.

🔄 chore(titipan.tsx): update dummy data for 'nama_titipan' and 'satuan' fields
The dummy data for the 'nama_titipan' field has been updated from 'Buku' to 'Kue Gacor' for more descriptive and accurate representation. Similarly, the 'satuan' field has been updated from 'Pieces' to 'Loyang' to provide clearer information about the unit of measurement.