🔧 fix(install.cmd): update references from 'produk' to 'product' in dummy data files 🔧 fix(admin): update references from 'produk' to 'product' in admin pages The changes were made to maintain consistency in naming conventions across the codebase by updating references from 'produk' to 'product' in dummy data files and admin pages.
🔧 fix(page.tsx): refactor code to use axios for API calls and improve code structure
🚀 feat(page.tsx): add functionality to fetch product types from API and handle form submission
The changes were made to refactor the code to use axios for making API calls, improve the structure of the code, and add functionality to fetch product types from the API. Additionally, the code now handles form submission to add a new product with the necessary data.

🔄 refactor(hampers.tsx): update import path from './produk' to './product'
🆕 feat(product.tsx): add Product interface and produk_data array for product details
🆕 feat(product_type.tsx): add ProductType interface for product type details
🗑️ chore(produk.tsx): remove outdated Produk interface and produk_data array
The import path in 'hampers.tsx' is updated to point to the correct file. New interfaces for Product and ProductType are added in separate files for better organization. The outdated Produk interface and produk_data array are removed as they are no longer needed.