🔄 chore(page.tsx): refactor code to use HampersFetch type instead of Hampers for currentHampers and updateHampers 🔧 fix(page.tsx): update fetchHampers function to correctly fetch and set currentHampers and updateHampers with data from API The changes were made to improve type consistency by using the HampersFetch type for the currentHampers and updateHampers states. The fetchHampers function was updated to correctly fetch data from the API, set the state values accordingly, and fetch the image associated with the hampers.
🔧 fix(page.tsx): remove unused satuanSelected state and related logic
🚀 feat(page.tsx): add functionality to fetch products from the server and update state
The commit removes the unused satuanSelected state and related logic to clean up the code. Additionally, it adds the functionality to fetch products from the server using an API call and update the state with the fetched data. This enhancement allows the page to dynamically load and display product data, improving the user experience.

🔧 fix(page.tsx): update data fetching logic to use axios for Hampers data
🚀 feat(page.tsx): add functionality to fetch and display images for each Hampers item
The changes were made to improve the data fetching process by switching to axios for fetching Hampers data. Additionally, the functionality to fetch and display images for each Hampers item was added to enhance the user experience.

🔨 fix(page.tsx): add Alert component from Material-UI to display success message
🚀 feat(page.tsx): implement functionality to display success message after adding hampers
The changes were made to improve user experience by providing visual feedback to the user after successfully adding a hampers item. The Alert component from Material-UI is used to display a success message when the hampers item is added successfully.

🚀 feat(admin): add new section for "Hampers" in the admin navbar
🔧 refactor(dummy_data): update data structures for hampers and products
The commit adds a new section for "Hampers" in the admin navbar, providing links to "Tambah Hampers" and "List Hampers". This enhances the navigation experience for users managing hampers in the admin panel. Additionally, the data structures for hampers and products have been refactored to include more detailed information and align with the application's requirements.