🔒 chore(layout.tsx): comment out router push to '/admin/produk/list' for debugging The router push to '/admin/produk/list' has been commented out for debugging purposes. This change allows for easier troubleshooting of the authentication flow without automatically redirecting to the specified route.
🔨 refactor(page.tsx): rename List component to ListResep for clarity and consistency
🚀 feat(page.tsx): implement search functionality to fetch and display search results
The component List is renamed to ListResep for better clarity and consistency in naming conventions. The search functionality is implemented to fetch search results from the server based on the search query entered by the user. This enhancement allows users to search for specific items efficiently within the application.

🔨 refactor(page.tsx): refactor delete confirmation dialog for better readability and maintainability
🚀 feat(page.tsx): rename export from List to ListResep for clarity and consistency
The changes were made to improve the structure and readability of the delete confirmation dialog code. By refactoring the code, it is now easier to understand and maintain. Additionally, renaming the export from List to ListResep enhances clarity and consistency within the codebase.

🔨 refactor(page.tsx): clean up unused code and fix formatting issues
🚀 feat(page.tsx): add functionality to select a product and update the product_id in the resep state
The changes involve cleaning up unused code and fixing formatting issues in the file. Additionally, a feature was added to allow users to select a product from a list and update the product_id in the resep state accordingly. This enhances the user experience and data accuracy in the application.

🔨 refactor(page.tsx): remove unnecessary code related to editing and deleting recipes
🧹 clean up the code by removing unused form elements, dialogs, and transitions related to editing and deleting recipes. This improves code readability and reduces complexity by eliminating redundant and unused functionality.