🔨 refactor(CakeHome.tsx): remove redundant code related to 'opsi_pengambilan' dropdown The code related to the 'opsi_pengambilan' dropdown has been removed as it was redundant and unnecessary. This cleanup improves the codebase by eliminating duplicate code and simplifying the component structure.
🔨 refactor(CartPage.tsx): remove unused imports and variables, add new state variables for delivery options and recipient details
🚀 feat(CartPage.tsx): implement state variables for delivery options and recipient details to enhance user experience
The commit refactors the CartPage component by removing unused imports and variables for better code cleanliness. It also introduces new state variables to manage delivery options and recipient details, improving the user experience by allowing them to select delivery options and provide recipient information.

🚀 feat(CartPage.tsx): add handleCheckout function to process checkout flow
🛒 fix(CartPage.tsx): adjust styling for text wrapping in product descriptions
The handleCheckout function is added to handle the checkout process when the user clicks the "Check out" button. It retrieves user data, calculates the total price, generates an invoice number, creates a new transaction object, and saves the transaction and its details. Additionally, styling adjustments are made to ensure proper text wrapping in product descriptions for better readability.

🔨 refactor(HampersHome.tsx): remove redundant code related to 'opsi_pengambilan' dropdown
The code related to the 'opsi_pengambilan' dropdown has been removed as it was redundant and unnecessary. This cleanup improves code readability and reduces complexity by eliminating duplicate code blocks that serve no functional purpose.

🔥 refactor(MinumanHome.tsx): remove redundant code related to 'opsi_pengambilan' dropdown
The code related to the 'opsi_pengambilan' dropdown has been removed as it was redundant and not being used in the application. This cleanup improves code readability and reduces unnecessary complexity.

🔥 refactor(ProdukHome.tsx): remove redundant code related to 'opsi_pengambilan' in ProdukHome component
🔥 refactor(ShippingAddress.tsx): delete ShippingAddress component as it is no longer needed
The changes were made to clean up the codebase by removing redundant code related to 'opsi_pengambilan' in the ProdukHome component and deleting the ShippingAddress component as it is no longer used in the application.

🔨 refactor(TitipanHome.tsx): remove unused opsi_pengambilan property in newCart object
🔨 refactor(cart.tsx): remove opsi_pengambilan property from cart_data objects
🔨 refactor(transaction.tsx): remove poin_user property and fix typo in total_poin_user property
🔨 refactor(transaction_detaill.tsx): fix typo in transaction detail properties
The changes were made to clean up the codebase by removing unused or redundant properties and fixing a typo in one of the property names for consistency and clarity.