🚀 chore(package.json): update dependencies versions and add new packages 🛠️ refactor(navbar.tsx): remove unnecessary code and add new state for quota 🔧 fix(CakeHome.tsx): fix quota display issue and add support for fetching quota The commit updates dependencies versions in package.json, adds new packages like react-to-pdf and react-to-print. In navbar.tsx, unnecessary code is removed, and a new state for quota is added. In CakeHome.tsx, the issue with quota display is fixed, and support for fetching quota data is added.
🔧 fix(CartPage.tsx): add missing import statement for User in CartPage.tsx
🔧 fix(HampersHome.tsx): add missing import statement for getToday function in HampersHome.tsx
🚀 feat(CartPage.tsx): implement functionality to fetch user data and update points
🚀 feat(HampersHome.tsx): implement functionality to fetch and display daily quota for hampers
The changes were made to CartPage.tsx to fetch user data and update points based on the user's actions during the transaction process. In HampersHome.tsx, the functionality to fetch and display the daily quota for hampers was implemented to provide users with real-time information on product availability.

🔨 refactor(Invoice.tsx): refactor Invoice component to improve readability and maintainability
🚀 feat(Invoice.tsx): add support for printing functionality using ReactToPrint library
The changes involve refactoring the Invoice component to improve code readability and maintainability. Additionally, support for printing functionality using the ReactToPrint library has been added, allowing users to easily print the invoice content.

🔨 refactor(MinumanHome.tsx): add getToday function to calculate current date
🔧 chore(MinumanHome.tsx): refactor variable names for clarity and consistency
The changes were made to improve code readability and maintainability by introducing a function to calculate the current date in the MinumanHome component. Variable names were also updated for better clarity and consistency in the codebase.

🔧 fix(TransaksiContent.tsx): fix import statement for FaArrowCircleLeft icon
🔧 fix(TransaksiContent.tsx): change role_id from 2 to 1 for customer registration
🔧 fix(transaction.tsx): change field name 'tanggal_ambil' to 'tanggal_pemesanan'
🚀 feat(TransaksiContent.tsx): add functionality to fetch and display images for transactions
🚀 feat(TransaksiContent.tsx): add functionality to handle and display transaction details
🚀 feat(TransaksiContent.tsx): add support for fetching and displaying customer email
The changes were made to improve the user experience by enhancing the display of transaction details, handling images, and ensuring correct data is displayed during customer registration.